+++ zorka-agent/src/test/java/com/jitlogic/zorka/agent/test/spy/SpyInstanceIntegTest.java
@@ -31,7 +31,7 @@ public class SpyInstanceIntegTest extends ZorkaFixture {
     public void testTrivialMethodRun() throws Exception {
         SpyDefinition sdef = SpyDefinition.instrument().onSubmit(spy.tdiff("E0", "R0", "S0"))
                 .include(spy.byMethod(BytecodeInstrumentationFixture.TCLASS1, "trivialMethod"))
-                .onSubmit(spy.zorkaStats("test", "test:name=${shortClassName}", "stats", "${methodName}", "R0", "S0"));
+                .onSubmit(spy.zorkaStats("test", "test:name=${shortClassName}", "stats", "${methodName}", "S0"));
 
         spyInstance.add(sdef);